move Profiler and Graph to independent files
