+++ .config/mypy/mypy_enabled.txt
@@ -39,6 +39,8 @@ scapy/layers/l2.py
 #scapy/contrib/http2.py  # needs to be fixed
 scapy/contrib/roce.py
 scapy/contrib/automotive/enumerator.py
+scapy/contrib/automotive/graph.py
+scapy/contrib/automotive/profiler.py
 scapy/contrib/automotive/ecu.py
 scapy/contrib/automotive/uds_scan.py
 scapy/contrib/automotive/gm/gmlanscan.py
