+++ scapy/contrib/automotive/enumerator.py
@@ -6,7 +6,6 @@
 # scapy.contrib.description = AutomotiveTestCase and AutomotiveTestCaseExecutor base classes  # noqa: E501
 # scapy.contrib.status = loads
 
-import functools
 import time
 
 from collections import defaultdict, OrderedDict
@@ -14,6 +13,8 @@ from itertools import product
 
 from scapy.compat import Any, Union, List, Optional, Iterable, \
     Dict, Tuple, Set, Callable, Type, cast, FAKE_TYPING
+from scapy.contrib.automotive.graph import Graph, _Edge
+from scapy.contrib.automotive.profiler import Profiler, profile
 from scapy.error import Scapy_Exception, log_interactive
 from scapy.utils import make_lined_table, SingleConversationSocket, EDecimal
 import scapy.modules.six as six
@@ -58,73 +59,6 @@ else:
     _AutomotiveTestCaseFilteredScanResult = namedtuple(  # type: ignore
         "_AutomotiveTestCaseFilteredScanResult",
         ["state", "req", "resp", "req_ts", "resp_ts"])
-
-
-def profile(state, enum=None):
-    # type: (Any, Optional[Any]) -> Callable[[Callable[..., Any]], Callable[..., Any]]  # noqa: E501
-    def profile_decorator(func):
-        # type: (Callable[..., Any]) -> Callable[..., Any]
-        @functools.wraps(func)
-        def wrapper_profiled(*args, **kwargs):  # type: ignore
-            with Profiler(state, enum):
-                value = func(*args, **kwargs)
-            return value
-
-        return wrapper_profiled
-    return profile_decorator
-
-
-class Profiler:
-    # For the profiling we'll use the unix time
-    # candump is using the same and thus it would be matchable
-
-    # _candump_fmt_date = datetime.datetime.now().strftime("%Y-%m-%d_%H%M%S")
-    # _filename_milestone = "milestones-%s.csv" % _candump_fmt_date
-    # _filename_profiling = "profiling-%s.csv" % _candump_fmt_date
-    _filename_milestone = "milestones.csv"
-    _filename_profiling = "profiling.csv"
-
-    _first = True
-    enabled = True
-
-    def __init__(self, state, enum=None):
-        # type: (Any, Optional[Any]) -> None
-        # We use the csv format
-        # If len(p) > 1, it would contain a ','
-        # For example: "[1, 2, 3]"
-        # Thus we replace all ',' with a ';' to avoid this
-        self.state = str(state).replace(",", ";")
-        self.enum = str(enum or state).replace(",", ";")
-        self.start_time = time.time()
-
-        if Profiler._first and Profiler.enabled:
-            Profiler._first = False
-            with open(Profiler._filename_profiling, mode="a") as f:  # noqa: E501
-                # Writing header
-                # Not required for milestone file because this is parsed
-                # manually instead of using `read_csv` of plotly
-                f.write("state,enumerator,start,end\n")
-
-    @classmethod
-    def write_milestone(cls, name):
-        # type: (str) -> None
-        if not cls.enabled:
-            return
-        with open(cls._filename_milestone, mode="a") as f:
-            f.write("%s,%f\n" % (name, time.time()))
-
-    def __enter__(self):
-        # type: () -> Profiler
-        self.start_time = time.time()
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        # type: (Any, Any, Any) -> None
-        if not Profiler.enabled:
-            return
-        with open(Profiler._filename_profiling, mode="a") as f:
-            f.write("%s,%s,%f,%f\n" % (
-                self.state, self.enum, self.start_time, time.time()))
 
 
 class AutomotiveTestCaseExecutorConfiguration(object):
@@ -183,102 +117,6 @@ _CleanupCallable = Optional[_TransitionCallable]
 _TransitionTuple = Tuple[_TransitionCallable, _CleanupCallable]
 
 
-class Graph(object):
-    def __init__(self):
-        # type: () -> None
-        """
-        self.edges is a dict of all possible next nodes
-        e.g. {'X': ['A', 'B', 'C', 'E'], ...}
-        self.__transition_functions has all the transition_functions
-        between two nodes, with the two nodes as a tuple as the key
-        e.g. {('X', 'A'): 7, ('X', 'B'): 2, ...}
-        """
-        self.edges = defaultdict(list)  # type: Dict[EcuState, List[EcuState]]
-        self.__transition_functions = {}  # type: Dict[Tuple[EcuState, EcuState], Optional[_TransitionTuple]]  # noqa: E501
-
-    def __reduce__(self):  # type: ignore
-        f, t, d = super(Graph, self).__reduce__()  # type: ignore
-        try:
-            del d["_Graph__transition_functions"]
-        except KeyError:
-            pass
-        return f, t, d
-
-    def add_edge(self, from_node, to_node, transition_function=None):
-        # type: (EcuState, EcuState, Optional[_TransitionTuple]) -> None  # noqa: E501
-        """
-        Inserts new edge in directional graph
-        :param from_node: start node of edge
-        :param to_node: end node of edge
-        :param transition_function: tuple with enter and cleanup function
-        """
-        Profiler.write_milestone(repr(to_node))
-        self.edges[from_node].append(to_node)
-        self.__transition_functions[(from_node, to_node)] = transition_function
-
-    def get_transition_function(self, from_node, to_node):
-        # type: (EcuState, EcuState) -> Optional[_TransitionTuple]  # noqa: E501
-        try:
-            return self.__transition_functions[(from_node, to_node)]
-        except KeyError:
-            return None
-
-    @property
-    def transition_functions(self):
-        # type: () -> Dict[Tuple[EcuState, EcuState], Optional[_TransitionTuple]]  # noqa: E501
-        return self.__transition_functions
-
-    @property
-    def nodes(self):
-        # type: () -> Union[List[EcuState], Set[EcuState]]
-        return set([n for _, p in self.edges.items() for n in p])
-
-    @staticmethod
-    def dijkstra(graph, initial, end):
-        # type: (Graph, EcuState, EcuState) -> List[EcuState]
-        """
-        Compute shortest paths from initial to end in graph
-        From https://benalexkeen.com/implementing-djikstras-shortest-path-algorithm-with-python/  # noqa: E501
-        :param graph: Graph where path is computed
-        :param initial: Start node
-        :param end: End node
-        :return: A path as list of nodes
-        """
-        shortest_paths = {initial: (None, 0)}  # type: Dict[EcuState, Tuple[Optional[EcuState], int]]  # noqa: E501
-        current_node = initial
-        visited = set()
-
-        while current_node != end:
-            visited.add(current_node)
-            destinations = graph.edges[current_node]
-            weight_to_current_node = shortest_paths[current_node][1]
-
-            for next_node in destinations:
-                weight = 1 + weight_to_current_node
-                if next_node not in shortest_paths:
-                    shortest_paths[next_node] = (current_node, weight)
-                else:
-                    current_shortest_weight = shortest_paths[next_node][1]
-                    if current_shortest_weight > weight:
-                        shortest_paths[next_node] = (current_node, weight)
-
-            next_destinations = {node: shortest_paths[node] for node in
-                                 shortest_paths if node not in visited}
-            if not next_destinations:
-                return []
-            # next node is the destination with the lowest weight
-            current_node = min(next_destinations,
-                               key=lambda k: next_destinations[k][1])
-
-        # Work back through destinations in shortest path
-        path = []
-        while current_node is not None:
-            path.append(current_node)
-            next_node = cast(EcuState, shortest_paths[current_node][0])
-            current_node = next_node
-        # Reverse path
-        path.reverse()
-        return path
 
 
 class AutomotiveTestCaseABC(ABC):
